+++ src/app/demo/demo-page.component.ts
@@ -4,6 +4,7 @@ import { HeaderComponent } from './layout/header/header.component';
 import { FooterComponent } from './layout/footer/footer.component';
 import { TableDemoComponent } from './sections/table/table-demo.component';
 import { ToastDemoComponent } from './sections/toast/toast-demo.component';
+import { SelectDemoComponent } from './sections/select/select-demo.component';
 import { StepperDemoComponent } from './sections/stepper/stepper-demo.component';
 import { PaginationDemoComponent } from './sections/pagination/pagination-demo.component';
 import { ToastContainerComponent } from '../shared/components/toast/toast-container.component';
@@ -11,7 +12,16 @@ import { ToastContainerComponent } from '../shared/components/toast/toast-contai
 @Component({
   selector: 'app-demo-page',
   standalone: true,
-  imports: [HeaderComponent, FooterComponent, StepperDemoComponent, TableDemoComponent, PaginationDemoComponent, ToastContainerComponent, ToastDemoComponent],
+  imports: [
+    HeaderComponent,
+    FooterComponent,
+    TableDemoComponent,
+    ToastDemoComponent,
+    SelectDemoComponent,
+    StepperDemoComponent,
+    PaginationDemoComponent,
+    ToastContainerComponent,
+  ],
   templateUrl: './demo-page.component.html',
   styleUrl: './demo-page.component.sass'
 })