+++ src/app/shared/i18n/dict/index.ts
@@ -1,9 +1,10 @@
-import { COMMON } from './common';
-import { SECTIONS } from './sections';
-import { STEPPER } from './stepper';
 import { TABLE } from './table';
-import { PAGINATION } from './pagination';
 import { TOAST } from './toast';
+import { COMMON } from './common';
+import { SELECT } from './select';
+import { STEPPER } from './stepper';
+import { SECTIONS } from './sections';
+import { PAGINATION } from './pagination';
 
 import { Dict } from '../../model/type/dict.type';
 
@@ -14,4 +15,5 @@ export const DICT: Dict = {
   ...TABLE,
   ...PAGINATION,
   ...TOAST,
+  ...SELECT,
 };